Refactor real time clock libary and add thread-safety #10.
